Refactored the code for voter's registeration
